Read a parse state of the lookahead test as a 32-bit state id
The runtime of the fork has 32-bit state ids since tree-sitter-cpp commit 5882494, and
`Language::lookahead_iterator` and `LookaheadIterator::reset_state` take a `u32`.
`test_lookahead_iterator_exhaustion` converted each state to `u16`, and the test module of the CLI
did not compile: two errors E0308 at language_test.rs lines 76 and 91. The test now converts each
state to `u32`. Its assertions do not change.

This commit ports a test of the tree-sitter repository to the runtime of tree-sitter-cpp, and no commit of tree-sitter-cpp is its source.

+++ crates/cli/src/tests/language_test.rs
@@ -72,7 +72,7 @@ fn test_lookahead_iterator_exhaustion() {
     let language = get_language("json");
 
     for state in 0..language.parse_state_count() {
-        let state = u16::try_from(state).unwrap();
+        let state = u32::try_from(state).unwrap();
         let mut lookahead = language.lookahead_iterator(state).unwrap();
 
         // A fresh iterator is not positioned on a symbol.
